Move script data to provider
