+++ provider/testdata/data/baked/props/scx_v1.rs
@@ -4,7 +4,7 @@ type DataStruct = < :: icu_properties :: provider :: ScriptWithExtensionsPropert
 pub static DATA: litemap::LiteMap<&str, &DataStruct, &[(&str, &DataStruct)]> =
     litemap::LiteMap::from_sorted_store_unchecked(&[("und", UND)]);
 static UND: &DataStruct = &::icu_properties::provider::ScriptWithExtensionsPropertyV1 {
-    data: ::icu_properties::script::ScriptWithExtensions {
+    data: ::icu_properties::provider::ScriptWithExtensions {
         trie: ::icu_collections::codepointtrie::CodePointTrie::from_parts(
             ::icu_collections::codepointtrie::CodePointTrieHeader {
                 high_start: 918016u32,
